+++ packages/nidle-chain/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.3](https://github.com/yanuoda/nidle/compare/v0.1.2...v0.1.3) (2022-02-10)
+
+
+### Bug Fixes
+
+* output merge ([20a4c11](https://github.com/yanuoda/nidle/commit/20a4c11c34c044eea0149dbdf6a8650d14d7e870))
+
+
+
 ## [0.1.2](https://github.com/yanuoda/nidle/compare/v0.1.1...v0.1.2) (2022-02-08)