+++ packages/nidle-plugin-merge/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.3](https://github.com/yanuoda/nidle/compare/v0.1.2...v0.1.3) (2022-02-10)
+
+
+
 ## [0.1.2](https://github.com/yanuoda/nidle/compare/v0.1.1...v0.1.2) (2022-02-08)